+++ ArcGISRuntimeSDKQt_CppSamples/Maps/DisplayDeviceLocationWithNmeaDataSources/displayDeviceLocationWithNmeaDataSources.pro
