+++ ArcGISRuntimeSDKQt_CppSamples/Maps/DisplayDeviceLocationWithNmeaDataSources/main.cpp
@@ -11,19 +11,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <QSettings>
-#include <QGuiApplication>
-#include <QQuickView>
-#include <QCommandLineParser>
+#ifdef Q_OS_WIN
+#include <Windows.h>
+#endif
+
+#include "DisplayDeviceLocationWithNmeaDataSources.h"
+#include "ArcGISRuntimeEnvironment.h"
+
 #include <QDir>
-#include <QQmlEngine>
+#include <QGuiApplication>
+#include <QQmlApplicationEngine>
 
 int main(int argc, char *argv[])
 {
   QGuiApplication::setAttribute(Qt::AA_EnableHighDpiScaling);
   QGuiApplication app(argc, argv);
-  app.setApplicationName(QStringLiteral("DisplayDeviceLocationWithNmeaDataSources - QML"));
-
+  app.setApplicationName(QStringLiteral("DisplayDeviceLocationWithNmeaDataSources - C++"));
 
   // Use of Esri location services, including basemaps and geocoding,
   // requires authentication using either an ArcGIS identity or an API Key.
@@ -35,29 +38,28 @@ int main(int argc, char *argv[])
   const QString apiKey = QStringLiteral("");
   if (apiKey.isEmpty())
   {
-    qWarning() << "Use of Esri location services, including basemaps, requires"
-                  "you to authenticate with an ArcGIS identity or set the API Key property.";
+      qWarning() << "Use of Esri location services, including basemaps, requires "
+                    "you to authenticate with an ArcGIS identity or set the API Key property.";
   }
   else
   {
-    QCoreApplication::instance()->setProperty("Esri.ArcGISRuntime.apiKey", apiKey);
+      Esri::ArcGISRuntime::ArcGISRuntimeEnvironment::setApiKey(apiKey);
   }
 
-  // Initialize application view
-  QQuickView view;
-  view.setResizeMode(QQuickView::SizeRootObjectToView);
+  // Initialize the sample
+  DisplayDeviceLocationWithNmeaDataSources::init();
 
+  // Initialize application view
+  QQmlApplicationEngine engine;
   // Add the import Path
-  view.engine()->addImportPath(QDir(QCoreApplication::applicationDirPath()).filePath("qml"));
+  engine.addImportPath(QDir(QCoreApplication::applicationDirPath()).filePath("qml"));
 
 #ifdef ARCGIS_RUNTIME_IMPORT_PATH_2
-  view.engine()->addImportPath(ARCGIS_RUNTIME_IMPORT_PATH_2);
+  engine.addImportPath(ARCGIS_RUNTIME_IMPORT_PATH_2);
 #endif
 
   // Set the source
-  view.setSource(QUrl("qrc:/Samples/DisplayInformation/DisplayDeviceLocationWithNmeaDataSources/DisplayDeviceLocationWithNmeaDataSources.qml"));
-
-  view.show();
+  engine.load(QUrl("qrc:/Samples/Maps/DisplayDeviceLocationWithNmeaDataSources/main.qml"));
 
   return app.exec();
 }
