+++ ArcGISRuntimeSDKQt_QMLSamples/Maps/DisplayDeviceLocationWithNmeaDataSources/displayDeviceLocationWithNmeaDataSources.pro
